requirements.txt
Django==1.7.1
djangorestframework==3.0.2
django-filter==0.9.1
Markdown==2.5.2
psycopg2==2.5.4
